+++ NEWS.rst
@@ -28,6 +28,7 @@ Version 0.8.9 - dev
     - NEW: BlockLayout.is_layout_block, True if block is a model space or paper space block definition
     - NEW: ezdxf.algebra.Arc helper class to create arcs from 2 points and an angle or radius, or from 3 points
     - NEW: ezdxf.algebra.Arc.add_to_layout() with UCS support to create 3D arcs
+    - NEW: rename paper space layouts by Drawing.layouts.rename(old_name, new_name)
     - CHANGE: Drawing.blocks.delete_block(name, safe=True), new parameter save, check if block is still referenced
       (raises DXFValueError)
     - CHANGE: Drawing.blocks.delete_all_blocks(safe=True), new parameter save, ignores blocks still referenced if safe is True